Make new selector private
