+++ packages/block-editor/src/store/selectors.js
@@ -43,6 +43,7 @@ import {
 	isSectionBlock,
 	getParentSectionBlock,
 	isZoomOut,
+	isContainerInsertableToInContentOnlyMode,
 } from './private-selectors';
 
 const { isContentBlock } = unlock( blocksPrivateApis );
@@ -1650,37 +1651,6 @@ const isBlockVisibleInTheInserter = (
 
 	return true;
 };
-
-/**
- * Determines if a container (clientId) allows insertion of blocks, considering contentOnly mode restrictions.
- *
- * @param {Object} state    Editor state.
- * @param {string} clientId The client ID of the container block.
- *
- * @return {boolean} Whether the container allows insertion.
- */
-export function isContainerInsertableToInContentOnlyMode( state, clientId ) {
-	const blockName = getBlockName( state, clientId );
-	const blockType = getBlockType( blockName );
-
-	// Look at the `blockType.allowedBlocks` field to determine whether this has limited allowed blocks.
-	const hasAllowedBlockList = blockType?.allowedBlocks?.length > 0;
-	const isContainerContentBlock = isContentBlock( blockName );
-	const isRootBlockMain = getSectionRootClientId( state ) === clientId;
-
-	// In contentOnly mode, containers shouldn't be inserted into unless:
-	// 1. they are a section root;
-	// 2. they are a content block with limited allowed blocks (e.g. list).
-	if (
-		! isRootBlockMain &&
-		( ( isContainerContentBlock && ! hasAllowedBlockList ) ||
-			! isContainerContentBlock )
-	) {
-		return false;
-	}
-
-	return true;
-}
 
 /**
  * Determines if the given block type is allowed to be inserted into the block list.
